adding tests to see how constructor resolution works.
In fact, it's not easy to detect primary constructors for a class
where several constructors are. The constructor may have
INVOKE_SPECIAL on the same base class, which is not the base class
at all. We should use more JVM specs to figure that our (or probably
better to let the uses select constructors they'd need)

Examples:

class X :Base() { constructor() {
  Base() // this is new Base() created, but InvokeSpecial is used for <init>
  X() // is it a call to base constructor or new X object creation?
  Base(X()) //which is the new instance, and which is the super call?
 }
}

All these examples are solvable if we'd track the JVM stack, e.g. if we would know what `this` is on the stack top to be sent to the INVOKE_SPECIAL instruction. That analysis is probably too trick (is it equivalent to the Turing Machine stop problem?)

 Well, we do not use the fact the super/this constructor call should happen closer to the first instruction